server/agui: support aggregator to aggregate AGUI events (#797)
Previously, each AG-UI event needed to be stored in the session via the
session service. In streaming scenarios, the large number of AG-UI
events put excessive pressure on the session service. This PR provides
an aggregator for AG-UI events, which can aggregate consecutive
TextContent events, reducing the pressure on the session service. It
also provides scheduled aggregation to prevent issues caused by a lack
of new events for extended periods, which could prevent data from being
stored in the database.